+++ packages/core/lib/combo/logic/world-transform.ts
@@ -1207,6 +1207,16 @@ export class LogicPassWorldTransform {
       this.removeLocations(data);
     }
 
+    if(!settings.shuffleIciclesOot) {
+      const data = POOL.oot.filter((x: any) => x.type === 'icicle').map((x: any) => gameId('oot', x.location, ' ')) as string[];
+      this.removeLocations(data);
+    }
+
+    if(!settings.shuffleIciclesMm) {
+      const data = POOL.mm.filter((x: any) => x.type === 'icicle').map((x: any) => gameId('mm', x.location, ' ')) as string[];
+      this.removeLocations(data);
+    }
+
     /* Carpenters */
     if (['open', 'single'].includes(settings.gerudoFortress)) {
       this.removeLocations(['OOT Gerudo Fortress Jail 2', 'OOT Gerudo Fortress Jail 3', 'OOT Gerudo Fortress Jail 4']);