+++ packages/core/lib/combo/settings/random.ts
@@ -802,6 +802,8 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesMm = true;
       base.shuffleRedBouldersOot = true;
       base.shuffleRedBouldersMm = true;
+      base.shuffleIciclesOot = true;
+      base.shuffleIciclesMm = true;
       break;
     case 3:
       base.shufflePotsOot = 'overworld';
@@ -824,6 +826,8 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesMm = true;
       base.shuffleRedBouldersOot = true;
       base.shuffleRedBouldersMm = true;
+      base.shuffleIciclesOot = true;
+      base.shuffleIciclesMm = true;
       break;
     case 4:
       base.shufflePotsOot = 'dungeons';
@@ -862,6 +866,8 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesMm = booleanWeighted(random, 0.5);
       base.shuffleRedBouldersOot = booleanWeighted(random, 0.5);
       base.shuffleRedBouldersMm = booleanWeighted(random, 0.5);
+      base.shuffleIciclesOot = booleanWeighted(random, 0.5);
+      base.shuffleIciclesMm = booleanWeighted(random, 0.5);
       break;
     }
   }