+++ packages/core/src/common/ovl/actors/Bg_Icicle/Bg_Icicle.h
@@ -26,6 +26,9 @@ typedef struct Actor_BgIcicle {
 #endif
     s16 shiverTimer;
     ColliderCylinder collider;
+
+    Xflag xflag;
+    u8 isExtended;
 } Actor_BgIcicle;
 
 #endif