+++ packages/core/tests/seeds/extrasanity-mq.test.ts
@@ -51,6 +51,8 @@ test("Can make a seed - MQ Extrasanity", async () => {
     fairySpotShuffleOot: true,
     shuffleRedBouldersOot: true,
     shuffleRedBouldersMm: true,
+    shuffleIciclesOot: true,
+    shuffleIciclesMm: true,
     mqDungeons: { type: 'all' },
   });
   expect(res).toBeDefined();